Observer Pattern applied to Package-DbAdapter relation
ince every package must send its content to some DbAdapter
in order to this perform the upsert operation on the Database
Package class was mixed with the ruby native Observer module.
So Every time that any package enrich its content, this will be
sent to the Adpater. The adapter will take care to open/close
sessions with the DB.